requirements.txt
psutil>=5.9.0
rich>=13.0.0
ping3>=4.0.0
dnspython>=2.4.0
requests>=2.31.0
pyyaml>=6.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
